# Improve RegexCompiler's ToLower perf