Profiling shows that `RegexOptions.Compiled | RegexOptions.IgnoreCase` regexes are spending a non-trivial amount of time invoking `char.ToLower(c, _culture)`, and a good chunk of this is just in making the virtual call to `_culture.TextInfo` inside of `char.ToLower`.  Instead of caching a `CultureInfo` and calling `char.ToLower(c, _culture)`, `RegexCompiler` can cache the `TextInfo` once and then call `_textInfo.ToLower(c)`, saving a virtual call on every character comparison.

That's the second commit.  The first commit addresses why this change hasn't been done until now.  We'd previously made the same change in `RegexInterpreter`, but in `RegexCompiler`, we need a temporary in order to swap parameters to make the `ToLower` lower call (by the time `CallToLower` is invoked, the `char` is already on the evaluation stack).  So, the first commit fixes how `RegexCompiler` manages temporary locals.  Previously, the developer would declare a few temporary locals at the beginning of the method, and then manually figure out how to use them in a safe manner.  With this change, we instead maintain a pool of locals that can be rented from just where they're needed.  That makes it possible for a helper like `CallToLower` to just grab a local for the brief period it needs one.

This PR handles the `RegexOptions.IgnoreCase` scenario.  https://github.com/dotnet/runtime/pull/35194 covers the `RegexOptions.IgnoreCase | RegexOptions.InvariantCulture` scenario, as there `RegexCompiler` uses `ToLowerInvariant`.

Review without whitespace diffs: https://github.com/dotnet/runtime/pull/35203/files?w=1
cc: @GrabYourPitchforks, @eerhardt, @pgovind 

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Text.RegularExpressions;

public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssemblies(new[] { typeof(Program).Assembly }).Run(args);

    private readonly Regex _regex = new Regex("Hello, \\w+.", RegexOptions.Compiled | RegexOptions.IgnoreCase);
    private readonly string _input = new string("This is a test to see how well this does.  Hello, DotNetBot.");

    [Benchmark] public bool Test() => _regex.IsMatch(_input);
}
```

| Method |        Job |           Toolchain |     Mean |   Error |  StdDev | Ratio |
|------- |----------- |-------------------- |---------:|--------:|--------:|------:|
|   Test | Job-ABLHBF | \master\corerun.exe | 207.0 ns | 0.42 ns | 0.37 ns |  1.00 |
|   Test | Job-VWJGVI |     \pr\corerun.exe | 149.9 ns | 0.33 ns | 0.25 ns |  0.72 |