@GrabYourPitchforks 2020-04-20
@stephentoub We recently changed `TextInfo.ToUpper(char)` / `ToLower(char)` to be non-virtual. Is this something you could take advantage of in your code by emitting `call` instead of `callvirt`? Or do you need to retain compatibility with downlevel platforms?