@stephentoub 2020-04-21
> We recently changed TextInfo.ToUpper(char) / ToLower(char) to be non-virtual. Is this something you could take advantage of in your code by emitting call instead of callvirt? Or do you need to retain compatibility with downlevel platforms? (Edit: I suppose you might be relying on the implicit null check that callvirt performs.)

It'll still invoke it non-virtually, and potentially even inline it; it'll just include the null check if needed.  You think it's worthwhile changing it to be call instead?  Last I knew, in general the C# compiler emits callvirt even for non-virtual methods, except in some circumstances, such as with the ?. operator.

If we want to change it, I think we should change other uses of callvirt as well, e.g. various string methods are currently invoked with callvirt instead of call:
https://github.com/dotnet/runtime/blob/774daf4e97b858ec63c5d94e3cd25fa3f65d85ab/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexCompiler.cs#L528